requirements.txt
Django
pytz
djangorestframework
